refactor: service 폴더 구분
