+++ __tests__/WebLottoManagerTest.js
@@ -1,4 +1,4 @@
-import WebLottoManager from "../src/WebLottoManager";
+import WebLottoManager from "../src/service/WebLottoManager";
 
 describe("WebLottoManager 테스트", () => {
   const mockGenerator = () => [1, 2, 3, 4, 5, 6];
